api/src/main/java/org/example/api/Service/ApplyService.java
package org.example.api.Service;

import org.example.api.domain.Coupon;
import org.example.api.repository.CouponRepository;
import org.springframework.stereotype.Service;

@Service
public class ApplyService {

    private final CouponRepository couponRepository;

    public ApplyService(CouponRepository couponRepository) {
        this.couponRepository = couponRepository;
    }

    public void apply(Long userId) {
        long count = couponRepository.count();

        if(count > 30){
            return;
        }

        couponRepository.save(new Coupon(userId));

    }
}
